# public/doc/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Essential life API",
        "description": "REST API consumed by Essential Life IOS and Android app",
        "contact": {
            "email": "ibnaalberunyibrahim@gmail.com"
        },
        "version": "2.1.0"
    },
    "paths": {
...
        "/{lang}/v2.1/userGeneratedContent/{uuid}": {
            "get": {
                "tags": [
                    "V2.1-api-auth"
                ],
                "description": "Fetch the element by uuid",
                "operationId": "App\\Http\\Controllers\\Api\\v2_1\\UserGeneratedContent\\UserGeneratedContentController::show",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "language",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "sp",
                                "jp"
                            ]
                        }
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserGeneratedContentResource2.1"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "V2.1-api-auth"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\v2_1\\UserGeneratedContent\\UserGeneratedContentController::update",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "language",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "sp",
                                "jp"
                            ]
                        }
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "is_public",
                                    "content"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "is_public": {
                                        "type": "boolean"
                                    },
                                    "content": {
                                        "type": "array",
                                        "items": {}
                                    },
                                    "base64_image": {
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserGeneratedContentResource2.1"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "V2.1-api-auth"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\v2_1\\UserGeneratedContent\\UserGeneratedContentController::destroy",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "language",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "sp",
                                "jp"
                            ]
                        }
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/{lang}/v3.0/userGeneratedContent/{uuid}": {
            "get": {
                "tags": [
                    "V3.0-api-auth"
                ],
                "description": "Fetch the element by uuid",
                "operationId": "App\\Http\\Controllers\\Api\\v3_0\\UserGeneratedContent\\UserGeneratedContentController::show",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "language",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "sp",
                                "jp"
                            ]
                        }
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserGeneratedContentResource3.0"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Model missing"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "V3.0-api-auth"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\v3_0\\UserGeneratedContent\\UserGeneratedContentController::update",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "language",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "sp",
                                "jp"
                            ]
                        }
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "is_public": {
                                        "type": "boolean"
                                    },
                                    "content": {
                                        "type": "array",
                                        "items": {}
                                    },
                                    "base64_image": {
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserGeneratedContentResource3.0"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "V3.0-api-auth"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\v3_0\\UserGeneratedContent\\UserGeneratedContentController::destroy",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "language",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "sp",
                                "jp"
                            ]
                        }
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "404": {
                        "description": "Content missing"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/{langs}/v3.0/userGeneratedContent": {
            "post": {
                "tags": [
                    "V3.0-api-auth"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\v3_0\\UserGeneratedContent\\UserGeneratedContentController::store",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "language",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "sp",
                                "jp"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "uuid",
                                    "name",
                                    "type"
                                ],
                                "properties": {
                                    "uuid": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "is_public": {
                                        "type": "boolean"
                                    },
                                    "content": {
                                        "type": "array",
                                        "items": {}
                                    },
                                    "base64_image": {
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserGeneratedContentResource3.0"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserGeneratedContentResource2.1": {
                "title": "User Generated Content Resource V3.0",
                "description": "Individual ailment remedy resource response",
                "properties": {
                    "uuid": {
                        "description": "Resource UUID",
                        "type": "integer"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "public_uuid": {
                        "type": "integer"
                    },
                    "name": {
                        "description": "Resource name",
                        "type": "string"
                    },
                    "image_url": {
                        "description": "Resource image url",
                        "type": "string"
                    },
                    "type": {
                        "description": "Readable resource name",
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "is_public": {
                        "type": "boolean"
                    },
                    "content": {
                        "type": "array",
                        "items": {}
                    },
                    "rejection_reason_subject": {
                        "type": "string"
                    },
                    "rejection_reason_description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UserGeneratedContentResource3.0": {
                "title": "User Generated Content Resource V3.0",
                "description": "Individual ailment remedy resource response",
                "properties": {
                    "uuid": {
                        "description": "Resource UUID",
                        "type": "integer"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "public_uuid": {
                        "type": "integer"
                    },
                    "name": {
                        "description": "Resource name",
                        "type": "string"
                    },
                    "image_url": {
                        "description": "Resource image url",
                        "type": "string"
                    },
                    "type": {
                        "description": "Readable resource name",
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "is_public": {
                        "type": "boolean"
                    },
                    "content": {
                        "type": "array",
                        "items": {}
                    },
                    "rejection_reason_subject": {
                        "type": "string"
                    },
                    "rejection_reason_description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}